Fix Intl messages disappearing on save
