+++ web/src/i18n/index.ts
@@ -1,6 +1,2 @@
-export {
-	default,
-	useIntl,
-	type IntlKey,
-	type IntlMessages,
-} from "./IntlProvider";
+export { default } from "./IntlProvider";
+export { useIntl, type IntlKey, type IntlMessages } from "./IntlContext";
