--- web/src/i18n/translations/es.json
@@ -1,72 +1,0 @@
-{
-	"_LANDING_": "",
-	"comeToTheRightPlace": "boop",
-	"enterRoomCodeHere": "boop again",
-	"or": "or",
-	"newRoom": "Start a new room",
-	"roomDoesNotExist": "A room does not exist with that code. Try again.",
-	"tooManyAttempts": "Too many attempts. Please wait and try again later.",
-
-	"_HEADER_": "",
-	"home": "Home",
-	"toggleTheme": "Toggle Theme",
-
-	"_JOIN_ROOM_": "",
-	"name": "Name",
-	"done": "Done",
-	"enterAName": "Please enter a name.",
-	"nameTooLong": "Name too long. Must be no more than 10 characters.",
-
-	"_CREATE_ROOM_": "",
-	"yourName": "Your name",
-	"voterOptions": "What options would you like voters to have to choose from?",
-	"fibonacci": "Fibonacci",
-	"linear": "Linear",
-	"yesNo": "Yes/No",
-	"numberOfOptions": "Number of options (min: {min}, max: {max})",
-	"includeNoVote": "Include no-vote option?",
-	"yes": "Yes",
-	"no": "No",
-	"finalPreview": "Final preview",
-	"errorWithMsg": "Error: {msg}",
-
-	"_ROOM_": "",
-	"connecting": "Connecting...",
-	"checkingRoom": "Checking room...",
-	"copyCode": "Click to copy code.",
-
-	"_ROOM_VOTER_VIEW_": "",
-	"waitingForModerator": "Waiting for {moderatorName}...",
-	"optionSelected": "Got it! You can change your mind if you want. Otherwise sit tight.",
-	"makeSelection": "Make a selection",
-
-	"_ROOM_VOTER_TABLE_": "",
-	"voters": "Voters",
-	"voted": "Voted",
-	"confidence": "Confidence",
-	"waitingForSelection": "Waiting for selection.",
-	"highConfidence": "High confidence",
-	"mediumConfidence": "Medium confidence",
-	"lowConfidence": "Low confidence",
-	"votingInProgress": "Voting in progress",
-	"waitingToStartVoting": "Waiting to start voting",
-	"waitingForVoters": "Waiting for people to join",
-	"noVotes": "No votes were cast.",
-	"low": "Low",
-	"high": "High",
-	"mode": "Mode",
-	"didNotVote": "Did not vote",
-
-	"_ROOM_VOTER_TABLE_MODERATOR_OPTIONS": "",
-	"makeModerator": "Make moderator",
-	"kickVoter": "Kick voter",
-	"makeModeratorDesc": "Are you sure you want to remove *{name}* from the room?",
-	"kickVoterDesc": "Are you sure you want to make *{name}* the new moderator? This will turn you into a voter.",
-	"confirm": "Confirm",
-	"cancel": "Cancel",
-
-	"_ROOM_MODERATOR_VIEW_": "",
-	"resetAndStartVoting": "Reset Votes & Start Voting",
-	"startVoting": "Start Voting",
-	"stopVoting": "Stop Voting"
-}
